pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brault"
version = "0.1.0"
description = "A tool for backing up and restoring data from HashiCorp Vault"
readme = "file: README.md"
authors = [
    { name = "teddy0605", email = "me@teddy.website" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "certifi==2024.2.2",
    "cffi==1.16.0",
    "charset-normalizer==3.3.2",
    "cryptography==42.0.5",
    "hvac==2.1.0",
    "idna==3.6",
    "pycparser==2.21",
    "PyYAML==6.0.1",
    "requests==2.31.0",
    "urllib3==2.2.1",
]
requires-python = ">=3.6"

[project.urls]
homepage = "https://github.com/teddy0605/brault"
issues = "https://github.com/teddy0605/brault/issues"

[project.scripts]
brault = "brault.cli:main"
